backend/requirements.txt
Flask==3.0.2
flask-marshmallow==1.2.1
Flask-SQLAlchemy==3.1.1
marshmallow==3.21.1
marshmallow-sqlalchemy==1.0.0
packaging==24.0
PyMySQL==1.1.0
SQLAlchemy==2.0.29
python-dotenv==1.0.1
cryptography==42.0.5
python-dotenv